+++ docs/make.jl
@@ -26,6 +26,7 @@ makedocs(;
         "Attribute Aliases" => "aliases.md",
         "Themes" => "themes.md",
         "Recipe System" => "recipes.md",
+        "Observables.jl" => "observables.md",
         "API Reference" => "api.md",
     ],
     warnonly=[:missing_docs, :docs_block],